extracted validation to utils
